добавления API пагинации для пользователя

+++ src/main/java/com/javamentor/qa/platform/models/dto/UserDto.java
@@ -24,5 +24,5 @@ public class UserDto {
     @Schema(description = "город пользователя")
     private String city;
     @Schema(description = "репутация пользователя")
-    private Long reputation;
+    private int reputation;
 }
